+++ src/main/java/org/truffleruby/cext/CExtNodes.java
@@ -1239,24 +1239,33 @@ public abstract class CExtNodes {
 
     }
 
-    @CoreMethod(names = "rb_call_super_splatted", onSingleton = true, rest = true)
+    @CoreMethod(names = "rb_call_super_splatted", onSingleton = true, required = 1, rest = true)
     public abstract static class CallSuperNode extends CoreMethodArrayArgumentsNode {
 
         @Child private CallSuperMethodNode callSuperMethodNode = CallSuperMethodNode.create();
 
         @Specialization
-        Object callSuper(VirtualFrame frame, Object[] args,
+        Object callSuper(VirtualFrame frame, boolean givenKeywords, Object[] args,
                 @Cached MetaClassNode metaClassNode) {
             final Frame callingMethodFrame = findCallingMethodFrame();
             final InternalMethod callingMethod = RubyArguments.getMethod(callingMethodFrame);
             final Object callingSelf = RubyArguments.getSelf(callingMethodFrame);
+            final Object block = RubyArguments.getBlock(callingMethodFrame);
             final RubyClass callingMetaclass = metaClassNode.execute(this, callingSelf);
             final MethodLookupResult superMethodLookup = ModuleOperations
                     .lookupSuperMethod(callingMethod, callingMetaclass);
             final InternalMethod superMethod = superMethodLookup.getMethod();
+
+            final ArgumentsDescriptor descriptor;
+            if (givenKeywords) {
+                descriptor = KeywordArgumentsDescriptor.EMPTY;
+            } else {
+                descriptor = NoKeywordArgumentsDescriptor.INSTANCE;
+            }
+
             // This C API only passes positional arguments, but maybe it should be influenced by ruby2_keywords hashes?
             return callSuperMethodNode.execute(
-                    frame, callingSelf, superMethod, NoKeywordArgumentsDescriptor.INSTANCE, args, nil);
+                    frame, callingSelf, superMethod, descriptor, args, block);
         }
 
         @TruffleBoundary
@@ -1269,6 +1278,7 @@ public abstract class CExtNodes {
                 if (method == null) {
                     return null;
                 } else if (method.getName().equals(/* Truffle::CExt. */ "rb_call_super") ||
+                        method.getName().equals(/* Truffle::CExt. */ "rb_call_super_kw") ||
                         method.getName().equals(/* Truffle::Interop. */ "execute_without_conversion") ||
                         method.getName().equals(/* Truffle::CExt. */ "rb_call_super_splatted")) {
                     // TODO CS 11-Mar-17 must have a more precise check to skip these methods